// typechain-types/rif-relay-contracts/interfaces/IRelayHub.ts
export declare namespace IForwarder {
  export type DeployRequestStruct = {
    relayHub: AddressLike;
    from: AddressLike;
    to: AddressLike;
    tokenContract: AddressLike;
    recoverer: AddressLike;
    value: BigNumberish;
    nonce: BigNumberish;
    tokenAmount: BigNumberish;
    tokenGas: BigNumberish;
    validUntilTime: BigNumberish;
    index: BigNumberish;
    data: BytesLike;
  };

  export type DeployRequestStructOutput = [
    relayHub: string,
    from: string,
    to: string,
    tokenContract: string,
    recoverer: string,
    value: bigint,
    nonce: bigint,
    tokenAmount: bigint,
    tokenGas: bigint,
    validUntilTime: bigint,
    index: bigint,
    data: string
  ] & {
    relayHub: string;
    from: string;
    to: string;
    tokenContract: string;
    recoverer: string;
    value: bigint;
    nonce: bigint;
    tokenAmount: bigint;
    tokenGas: bigint;
    validUntilTime: bigint;
    index: bigint;
    data: string;
  };

  export type ForwardRequestStruct = {
    relayHub: AddressLike;
    from: AddressLike;
    to: AddressLike;
    tokenContract: AddressLike;
    value: BigNumberish;
    gas: BigNumberish;
    nonce: BigNumberish;
    tokenAmount: BigNumberish;
    tokenGas: BigNumberish;
    validUntilTime: BigNumberish;
    data: BytesLike;
  };

  export type ForwardRequestStructOutput = [
    relayHub: string,
    from: string,
    to: string,
    tokenContract: string,
    value: bigint,
    gas: bigint,
    nonce: bigint,
    tokenAmount: bigint,
    tokenGas: bigint,
    validUntilTime: bigint,
    data: string
  ] & {
    relayHub: string;
    from: string;
    to: string;
    tokenContract: string;
    value: bigint;
    gas: bigint;
    nonce: bigint;
    tokenAmount: bigint;
    tokenGas: bigint;
    validUntilTime: bigint;
    data: string;
  };
}

export declare namespace EnvelopingTypes {
  export type RelayDataStruct = {
    gasPrice: BigNumberish;
    feesReceiver: AddressLike;
    callForwarder: AddressLike;
    callVerifier: AddressLike;
  };

  export type RelayDataStructOutput = [
    gasPrice: bigint,
    feesReceiver: string,
    callForwarder: string,
    callVerifier: string
  ] & {
    gasPrice: bigint;
    feesReceiver: string;
    callForwarder: string;
    callVerifier: string;
  };

  export type DeployRequestStruct = {
    request: IForwarder.DeployRequestStruct;
    relayData: EnvelopingTypes.RelayDataStruct;
  };

  export type DeployRequestStructOutput = [
    request: IForwarder.DeployRequestStructOutput,
    relayData: EnvelopingTypes.RelayDataStructOutput
  ] & {
    request: IForwarder.DeployRequestStructOutput;
    relayData: EnvelopingTypes.RelayDataStructOutput;
  };

  export type RelayRequestStruct = {
    request: IForwarder.ForwardRequestStruct;
    relayData: EnvelopingTypes.RelayDataStruct;
  };

  export type RelayRequestStructOutput = [
    request: IForwarder.ForwardRequestStructOutput,
    relayData: EnvelopingTypes.RelayDataStructOutput
  ] & {
    request: IForwarder.ForwardRequestStructOutput;
    relayData: EnvelopingTypes.RelayDataStructOutput;
  };
}